requirements.txt
Flask>=3.1.0
gunicorn>=21.2.0
pandas>=2.2.0
numpy>=2.2.0
scikit-learn>=1.6.0
mysql-connector-python>=9.2.0
python-dotenv>=1.0.1
